[US339662] Add mas and mas-foundation module.

+++ mas-foundation/src/androidTest/java/com/ca/mas/core/storage/AccountManagerStorageTests.java
@@ -50,6 +50,7 @@ public class AccountManagerStorageTests extends BaseStorageTests {
     @Override
     @After
     public void tearDown() throws Exception {
+        if (isSkipped) return;
         currentStorage.deleteAll();
     }
